Core - Updated functionality in the Native namespace.
- Updated GorgonNativeBuffer to simplify its available functionality.
- Updated GorgonPtr to increase its functionality and usefulness.
- Updated GorgonNativeExtensions to be more representative of the design goals for the buffer and pointer types.
- Added tests for the native buffer, pointer type and extension methods.

+++ Gorgon/Gorgon.Graphics.Core/Textures/GorgonTexture1D.cs
@@ -1019,7 +1019,7 @@ public sealed class GorgonTexture1D
                                                                  Right = finalBounds.Right,
                                                                  Bottom = finalBounds.Bottom
                                                              },
-                                                             imageBuffer.Data,
+                                                             (nint)imageBuffer.Data,
                                                              imageBuffer.PitchInformation.RowPitch,
                                                              imageBuffer.PitchInformation.RowPitch,
                                                              (int)copyMode);
