+++ Gorgon/Gorgon.Graphics.Core/Textures/GorgonTexture2D.cs
@@ -1410,7 +1410,7 @@ public sealed class GorgonTexture2D
                                                                  Right = finalBounds.Right,
                                                                  Bottom = finalBounds.Bottom
                                                              },
-                                                             imageBuffer.Data,
+                                                             (nint)imageBuffer.Data,
                                                              imageBuffer.PitchInformation.RowPitch,
                                                              imageBuffer.PitchInformation.SlicePitch,
                                                              (int)copyMode);
