+++ Gorgon/Gorgon.Graphics.Core/Textures/GorgonTexture3D.cs
@@ -1148,7 +1148,7 @@ public sealed class GorgonTexture3D
                                                                  Right = finalBounds.Right,
                                                                  Bottom = finalBounds.Bottom
                                                              },
-                                                             imageBuffer.Data,
+                                                             (nint)imageBuffer.Data,
                                                              imageBuffer.PitchInformation.RowPitch,
                                                              imageBuffer.PitchInformation.SlicePitch,
                                                              (int)copyMode);
